bit manipulation

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -8,15 +8,15 @@
  *
  * Return: the number of bits to flip to get from n to m.
  */
-
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-        unsigned long int diff = n ^ m;
-        unsigned int count = 0;
-        while (diff)
-        {
-                count += diff & 1;
-                diff >>= 1;
-        }
-        return (count);
+	unsigned long int diff = n ^ m;
+	unsigned long int count = 0;
+
+	while (diff)
+	{
+		count += diff & 1;
+		diff >>= 1;
+	}
+	return (count);
 }
